Solid-Waffle-Beginner-JavaScript-Solutions
A simple function that iterates over a string of characters to detect the single instance of an apostrophe and count space characters. The loop will continue to iterate until an apostrophe is found. When the function is invoked, it will return the total count of characters that preceded the apostrophe. You can rewrite the function to detect any alphanumeric key or character in the Unicode character set.